Removing dotenv dependency for elite to work
Depending on dotenv does not work properly. For some reasons that I could not debug the dotenv.findenv(usecwd=True) function did not work properly. It worked in debug environment but it stopped working when I launched the application. I tried to debug but as it was a third-party dependency I was not able to debug it.

So I decided to simply add extra argument to each __init__ constructor.